requirements.txt
python-keystoneclient
python-glanceclient
python-novaclient
python-neutronclient
python-cinderclient
python-swiftclient